requirements.txt
flask
sentence-transformers
faiss-cpu
transformers
mysql-connector-python